pyproject.toml
[tool.poetry]
name = "crop-yield-datascience-fullstack"
version = "0.1.0"
description = "A fullstack datascience project about crop yield"
authors = ["Ronan <ronan.griot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.2.5"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
python-slugify = "^8.0.4"
SQLAlchemy = "^2.0.40"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.2"
python-dotenv = "^1.1.0"
geopandas = "^1.0.1"
GeoAlchemy2 = "^0.17.1"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
